requirements.txt
django>=3.2,<3.3.0
djangorestframework>=3.13.1,<3.14.0
psycopg2>=2.9.3,<2.10.0
django-rest-auth==0.9.5
